requirements.txt
transformers
peewee
python-telegram-bot
python-dotenv
requests
pyshorteners
openai
pymysql
flask
functools